Some tidy up and renaming of mock-sensor expect functions
