+++ integration-tests/suites/socat.go
@@ -65,7 +65,7 @@ func (s *SocatTestSuite) TearDownSuite() {
 
 func (s *SocatTestSuite) TestSocat() {
 	// processes := s.Sensor().Processes(s.serverContainer)
-	endpoints := s.Sensor().ExpectNEndpoints(s.T(), s.serverContainer, 10*time.Second, 2)
+	endpoints := s.Sensor().ExpectEndpointsN(s.T(), s.serverContainer, 10*time.Second, 2)
 
 	// assert.Equal(s.T(), 3, len(processes))
